requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
faiss-cpu
pandas
matplotlib==3.7.1
reportlab
opencv-python
numpy==1.24.4
pillow
python-dotenv
pytesseract
scikit-image
